+++ process_manager.py
@@ -43,7 +43,6 @@ class ProcessManager:
             # the queues are not there yet to send stop msgs
             for process in (
                         'ProfilerProcess',
-                        'logsProcess',
                         'OutputProcess'
 
             ):
@@ -54,9 +53,7 @@ class ProcessManager:
             stop_msg = 'stop_process'
             self.main.profilerProcessQueue.put(stop_msg)
             self.main.outputqueue.put(stop_msg)
-            if hasattr(self.main, 'logsProcessQueue'):
-                self.logsProcessQueue.put(stop_msg)
-    
+
     def get_modules(self, to_ignore):
         """
         Get modules from the 'modules' folder.