+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Fix how we detect vertical and horizontal portscans
 - Update whitelist by adding all the ips of whitelisted domains
 - Fixed error whitelisting Unencrypted HTTP traffic 
+- Remove the feature of creating log directories using -l, now the only logs Slips generates are store in the output/ dir
+
+
 -1.0.4 (April 2023)
 - Add more descriptive titles to VT scores in the web UI thanks to @shubhangi.
 - Add stratoletters documentation, thanks to @haleelsada.